Move parse_profile
