+++ abm/common.py
@@ -1,8 +1,12 @@
+import os
 import sys
+import yaml
 import bioblend.galaxy
 
 GALAXY_SERVER = None
 API_KEY = None
+PROFILE_SEARCH_PATH = ['~/.abm/profile.yml', '.abm-profile.yml']
+
 
 datasets = {
     "dna": [
@@ -35,3 +39,31 @@ def connect():
         print('  3. Pass the APi key with the -k|--key option')
         sys.exit(1)
     return bioblend.galaxy.GalaxyInstance(url=GALAXY_SERVER, key=API_KEY)
+
+
+def parse_profile(profile_name: str):
+    '''
+    Parse the profile containing Galaxy URLs and API keys.
+
+    :param profile_name: path to the profile to parse
+    :return: a tuple containing the Galaxy URL and API key
+    '''
+    profiles = None
+    for profile_path in PROFILE_SEARCH_PATH:
+        profile_path = os.path.expanduser(profile_path)
+        if os.path.exists(profile_path):
+            with open(profile_path, 'r') as f:
+                # print(f'Loading profile from {profile_path}')
+                profiles = yaml.safe_load(f)
+            break
+    if profiles is None:
+        print(f'ERROR: Could not locate an abm profile file in {PROFILE_SEARCH_PATH}')
+        return None, None
+    if profile_name not in profiles:
+        print(f'ERROR: {profile_name} is not the name of a valid profile.')
+        print(f'The defined profile names are {profiles.keys()}')
+        return None, None
+    profile = profiles[profile_name]
+    return (profile['url'], profile['key'])
+
+
